add protect each route.

+++ controllers/contactController.js
@@ -2,7 +2,7 @@ const asyncHandler = require("express-async-handler");
 const Contact = require("../models/contactModel");
 
 const getAllContacts = asyncHandler(async (req, res) => {
-  const contacts = await Contact.find({});
+  const contacts = await Contact.find({ user_id: req.user._id });
   res.status(200).json(contacts);
 });
 
@@ -13,7 +13,12 @@ const createContact = asyncHandler(async (req, res) => {
       message: "All fields are required",
     });
   }
-  const contact = await Contact.create({ name, email, phone });
+  const contact = await Contact.create({
+    name,
+    email,
+    phone,
+    user_id: req.user._id,
+  });
   res.status(201).json(contact);
 });
 
@@ -29,7 +34,18 @@ const getContact = asyncHandler(async (req, res) => {
 
 const updateContact = asyncHandler(async (req, res) => {
   const { id: contactId } = req.params;
-  const contact = await Contact.findByIdAndUpdate(contactId, req.body, {
+  const contact = await Contact.findById(contactId);
+  if (!contact) {
+    res.status(404);
+    throw new Error("Contact not found");
+  }
+
+  if (contact.user_id.toString() !== req.user._id.toString()) {
+    res.status(401);
+    throw new Error("You are not authorized to update this contact");
+  }
+
+  const updateContact = await Contact.findByIdAndUpdate(contactId, req.body, {
     new: true,
     runValidators: true,
   });
@@ -37,17 +53,28 @@ const updateContact = asyncHandler(async (req, res) => {
     res.status(404);
     throw new Error("Contact not found");
   }
-  res.status(200).json(contact);
+  res.status(200).json(updateContact);
 });
 
 const deleteContact = asyncHandler(async (req, res) => {
   const { id: contactId } = req.params;
-  const contact = await Contact.findByIdAndDelete(contactId);
+  const contact = await Contact.findById(contactId);
   if (!contact) {
     res.status(404);
     throw new Error("Contact not found");
   }
-  res.status(200).json(contact);
+
+  if (contact.user_id.toString() !== req.user._id.toString()) {
+    res.status(401);
+    throw new Error("You are not authorized to update this contact");
+  }
+
+  const deleteContact = await Contact.findByIdAndDelete(contactId);
+  if (!contact) {
+    res.status(404);
+    throw new Error("Contact not found");
+  }
+  res.status(200).json(deleteContact);
 });
 
 module.exports = {
